requirements.txt
aiosignal==1.3.1
analytics-python==1.4.post1
annotated-types==0.7.0
bioc==2.1
certifi==2024.6.2
charset-normalizer==3.3.2
docutils==0.20.1
environs==9.5.0
filelock==3.14.0
filetype==1.2.0
fsspec==2024.6.0
ftfy==6.2.0
grpcio==1.64.1
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
marker-pdf==0.2.13
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pdftext==0.3.10
pillow==10.3.0
pydub==0.25.1
pypdfium2==4.30.0
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==1.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.9.3
readme-renderer==41.0
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sniffio==1.3.0
surya-ocr==0.4.12
sympy==1.12.1
tabulate==0.9.0
texify==0.1.9
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
twine==4.0.2
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13